pyproject.toml
[project]
name = "podlake"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "marctable",
  "httpx",
  "typer>=0.16.0",
  "python-dotenv>=1.1.0",
  "humanize>=4.12.3",
  "boto3>=1.38.24",
  "sickle>=0.7.0",
  "tqdm>=4.0.0",
]

[dependency-groups]
dev = [
    "boto3-stubs[s3,sts]>=1.38.25",
    "lxml-stubs>=0.5.1",
    "moto[s3,sts]>=5.1.5",
    "pandas-stubs>=2.2.3.250527",
    "pytest>=8.3.4",
    "ruff>=0.9.6",
    "ty>=0.0.31",
]

[project.urls]
homepage = "https://github.com/sul-dlss-labs/marctable"

[project.scripts]
podlake = "podlake:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.sdist]
packages = ["src/podlake"]

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "WARNING"
pythonpath = [
    "src"
]

[tool.pyright]
venvPath = "."
venv = ".venv"
